requirements.txt
flatbuffers==1.12
numpy==1.19.3
Pillow==8.4.0
pyenet==1.3.14